vram

+++ diffsynth/models/z_image_dit.py
@@ -51,7 +51,7 @@ class TimestepEmbedder(nn.Module):
 
     def forward(self, t):
         t_freq = self.timestep_embedding(t, self.frequency_embedding_size)
-        t_emb = self.mlp(t_freq.to(self.mlp[0].weight.dtype))
+        t_emb = self.mlp(t_freq.to(torch.bfloat16))
         return t_emb
